+++ core/datastore/src/main/java/com/ohnalmwo/datastore/di/LocalDataSourceModule.kt
@@ -1,7 +1,9 @@
 package com.ohnalmwo.datastore.di
 
-import com.ohnalmwo.datastore.datasource.LocationsDataSource
-import com.ohnalmwo.datastore.datasource.LocationsDataSourceImpl
+import com.ohnalmwo.datastore.datasource.location.LocationsDataSource
+import com.ohnalmwo.datastore.datasource.location.LocationsDataSourceImpl
+import com.ohnalmwo.datastore.datasource.setting.SettingDataSource
+import com.ohnalmwo.datastore.datasource.setting.SettingDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -16,4 +18,10 @@ abstract class LocalDataSourceModule {
     abstract fun bindLocationsDataSource(
         locationsDataSourceImpl: LocationsDataSourceImpl
     ): LocationsDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindSettingDataSource(
+        settingDataSourceImpl: SettingDataSourceImpl
+    ): SettingDataSource
 }
